+++ Codex/src/main/java/org/operatorfoundation/codex/symbols/Symbol.kt
@@ -1,5 +1,7 @@
 package org.operatorfoundation.codex.symbols
 
+import java.math.BigInteger
+
 /**
  * Base interface for all symbol types that can encode/decode values.
  * Symbols represent different character sets or value ranges (e.g., letters, numbers, bytes).
@@ -21,7 +23,7 @@ interface Symbol
      * @param encodedValue Can be Char, String, or Int depending on the symbol
      * @return The decoded integer value
      */
-    fun decode(encodedValue: ByteArray): Int
+    fun decode(encodedValue: ByteArray): BigInteger
 
     /**
      * Encodes a numeric value to its symbol representation.
@@ -30,5 +32,5 @@ interface Symbol
      * @param numericValue The integer to encode
      * @return Can be Char, String, or Int depending on the symbol
      */
-    fun encode(numericValue: Int): ByteArray
+    fun encode(numericValue: BigInteger): ByteArray
 }
